+++ src/core/floor.rs
@@ -46,6 +46,7 @@ struct FloorDto {
     bounding_box: Option<crate::core::spatial::BoundingBox3D>,
     wings: Vec<Wing>,
     equipment: Vec<String>,
+    #[serde(default, with = "crate::utils::sorted_map")]
     properties: HashMap<String, String>,
     #[serde(skip_serializing_if = "Option::is_none", default)]
     ifc_global_id: Option<String>,